+++ CHANGELOG.md
@@ -1,7 +1,11 @@
+### [0.10.0](https://github.com/xdevguild/buildo-begins/releases/tag/v0.10.0) (2023-01-29)
+- new `buildo-begins init` functionality. For now, it lets to initialize an example smart contract and NextJS dapp template. There will be more in the future.
+- dependencies updates
+
 ### [0.9.0](https://github.com/xdevguild/buildo-begins/releases/tag/v0.9.0) (2023-01-14)
 - rebrand to multiversx (continuation)
     - npm packages are replaced
-    - public api/explorer endpoints are replaced
+    -public api/explorer endpoints are replaced
 - update dependencies
 
 ### [0.8.0](https://github.com/xdevguild/buildo-begins/releases/tag/v0.8.0) (2023-01-07)